refactor: extract viewport service
